refactor: [9] 코드 리팩토링
