+++ web/src/store/index.ts
@@ -1,12 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
-import loginReducer from './loginSlice';
+import userReducer from '@/store/user';
 
 const store = configureStore({
   reducer: {
-    login: loginReducer,
+    login: userReducer,
   },
 });
 
 export default store;
+
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
