+++ web/src/types/HomePage.d.ts
@@ -1,4 +1,4 @@
-export interface LoginState {
+export interface UserState {
   id: string;
   password: string;
 }
